Examples of the close/far angle system
Both cars in these diagrams would be on orange team trying to score on blue, or clear the ball from its half. Far away shots have wider angles that the car can hit them from (shown with the red lines) this should lead to  less time for it to get in a tighter range and more time just hitting the ball. However, as the ball gets closer to the enemy goal, the angle tightens for increased accuracy.